fix: multiple merge errors, and underscores in var names

+++ pkg/pool/delta.go
@@ -3,18 +3,18 @@ package pool
 type Delta struct {
 	// Create defines the number of instances to spin up in the pool.
 	// Value is positive.
-	Create int
+	Create int64
 
 	// Delete defines the number of instances to shoot out of the pool.
 	// Value is positive.
-	Delete int
+	Delete int64
 }
 
 // NewDelta computes the operations to be performed on a challenge update
 // that resizes the pool.
 //
 // Expects all integers to be positive.
-func NewDelta(newMin, newMax, numClaimed, numPooled int) (d Delta) {
+func NewDelta(newMin, newMax, numClaimed, numPooled int64) (d Delta) {
 	desiredPool := computeDesiredPool(newMin, newMax, numClaimed)
 
 	if desiredPool > numPooled {
@@ -26,7 +26,7 @@ func NewDelta(newMin, newMax, numClaimed, numPooled int) (d Delta) {
 	return
 }
 
-func computeDesiredPool(minVal, maxVal, claimed int) int {
+func computeDesiredPool(minVal, maxVal, claimed int64) int64 {
 	if maxVal == 0 {
 		return minVal
 	}
